requirements.txt
click==8.1.6
colorama==0.4.6
Flask==2.2.5
greenlet==2.0.2
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
numpy==1.21.6
pandas==1.3.5
protobuf==3.20.3
psycopg2-binary
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
